Move all errors to their own sub packages
